requirements.txt
certifi==2020.12.5
chardet==4.0.0
cmake==3.18.4.post1
dlib==19.21.1
enum34==1.1.6
idna==2.10
install==1.3.4
joblib==1.0.0
keyboard==0.13.5
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
Pillow==8.1.0
pinnacle==0.0.8
pygame==2.0.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
PySide6==6.0.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.13.0
scikit-learn==0.24.1
scipy==1.6.0
shiboken6==6.0.1
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
urllib3==1.26.2
